[GH-2087] Geopandas: Share geoseries functions in base class to implement them for geodataframe (#2163)
* Add geometryarray.py as a modified geoseries.py copy

* Example: delegate area() call to base.py

* clean up

* Fix to_geoframe and delegate union_all and crosses for more examples

* Move geometryarray.py to geopandas folder, make to_geoframe method public

* raise ValueError instead of AttributeError when active geom col not set

* Don't name the result of function calls

* Add geodataframe tests for area, union_all, and crosses inside test_geoseries

* Use new MissingGeometryColumnError and bug fixes

* Fix tests by keeping name for set_crs

* Keep name for fillna, add tests for checking name

* Delegate centroid to shared base class

* Delegate geom_type

* Delegate length, is_valid, is_valid_reason, and is_empty. Adding support for functions w/ no args. Delete gdf geom_type

* Delegate dwithin, difference, has_z, boundary, get_geometry, is_simple. Also fix test_difference bug

* Move envelope, make_valid, intersects, overlaps, touches, within, covers, covered_by, distance, intersection, and snap

* Move contains and fillna. Modify _delegate_property logic to support inplace for fillna. Delete _row_wise_operation in GeoSeries

* bounds, total_bounds, estimate_utm_crs

* from_file, to_file, to_json, from_arrow, to_arrow, to_parquet, to_wkt, to_wkb

* Remove from_* functions and _create_from_select from geometryarray.py

* notna, isna, and improvements to rename + clean up utility functions

* Delete docstrings from geometryarray

* buffer and sindex

* Checkout geoseries.py from master and copy changes over. _delegate_to_geometry_column in base.py. remove some docstrings from geoseries.py

* Delete docstrings in geoseries

* DELETE geometryarray.py and uses of it. Fix funciton name for _delegate function

* Properly delegate geodataframe.buffer(), remove non-implemented geoseries from geodataframe.py, remove all abstract method decorators in base.py

* Skip shapely < 2 in test_sindex.py

* Clean up

* Fix test_io test by checking for name 'geometry'

* Use spark df api instead of spark sql in _create_from_select
