+++ python/sedona/geopandas/geoseries.py
@@ -637,7 +637,7 @@ class GeoSeries(GeoFrame, pspd.Series):
         new_epsg = crs.to_epsg() if crs else 0
 
         spark_col = stf.ST_SetSRID(self.spark.column, new_epsg)
-        result = self._query_geometry_column(spark_col)
+        result = self._query_geometry_column(spark_col, keep_name=True)
 
         if inplace:
             self._update_inplace(result)
@@ -655,6 +655,7 @@ class GeoSeries(GeoFrame, pspd.Series):
         df: pyspark.sql.DataFrame = None,
         returns_geom: bool = True,
         is_aggr: bool = False,
+        keep_name: bool = False,
     ) -> Union["GeoSeries", pspd.Series]:
         """
         Helper method to query a single geometry column with a specified operation.
@@ -678,7 +679,10 @@ class GeoSeries(GeoFrame, pspd.Series):
 
         df = self._internal.spark_frame if df is None else df
 
-        rename = self.name if self.name else SPARK_DEFAULT_SERIES_NAME
+        rename = SPARK_DEFAULT_SERIES_NAME
+
+        if keep_name and self.name:
+            rename = self.name
 
         col_expr = spark_col.alias(rename)
 
@@ -815,26 +819,6 @@ class GeoSeries(GeoFrame, pspd.Series):
 
     @property
     def area(self) -> pspd.Series:
-        """
-        Returns a Series containing the area of each geometry in the GeoSeries expressed in the units of the CRS.
-
-        Returns
-        -------
-        Series
-            A Series containing the area of each geometry.
-
-        Examples
-        --------
-        >>> from shapely.geometry import Polygon
-        >>> from sedona.geopandas import GeoSeries
-
-        >>> gs = GeoSeries([Polygon([(0, 0), (1, 0), (1, 1), (0, 1)]), Polygon([(0, 0), (2, 0), (2, 2), (0, 2)])])
-        >>> gs.area
-        0    1.0
-        1    4.0
-        dtype: float64
-        """
-
         spark_col = stf.ST_Area(self.spark.column)
 
         return self._query_geometry_column(
@@ -844,27 +828,6 @@ class GeoSeries(GeoFrame, pspd.Series):
 
     @property
     def geom_type(self) -> pspd.Series:
-        """
-        Returns a series of strings specifying the geometry type of each geometry of each object.
-
-        Note: Unlike Geopandas, Sedona returns LineString instead of LinearRing.
-
-        Returns
-        -------
-        Series
-            A Series containing the geometry type of each geometry.
-
-        Examples
-        --------
-        >>> from shapely.geometry import Polygon, Point
-        >>> from sedona.geopandas import GeoSeries
-
-        >>> gs = GeoSeries([Polygon([(0, 0), (1, 0), (1, 1), (0, 1)]), Point(0, 0)])
-        >>> gs.geom_type
-        0    POLYGON
-        1    POINT
-        dtype: object
-        """
         spark_col = stf.GeometryType(self.spark.column)
         result = self._query_geometry_column(
             spark_col,
@@ -893,32 +856,6 @@ class GeoSeries(GeoFrame, pspd.Series):
 
     @property
     def length(self) -> pspd.Series:
-        """
-        Returns a Series containing the length of each geometry in the GeoSeries.
-
-        In the case of a (Multi)Polygon it measures the length of its exterior (i.e. perimeter).
-
-        For a GeometryCollection it measures sums the values for each of the individual geometries.
-
-        Returns
-        -------
-        Series
-            A Series containing the length of each geometry.
-
-        Examples
-        --------
-        >>> from shapely.geometry import Polygon
-        >>> from sedona.geopandas import GeoSeries
-
-        >>> gs = GeoSeries([Point(0, 0), LineString([(0, 0), (1, 1)]), Polygon([(0, 0), (1, 0), (1, 1)]), GeometryCollection([Point(0, 0), LineString([(0, 0), (1, 1)]), Polygon([(0, 0), (1, 0), (1, 1)])])])
-        >>> gs.length
-        0    0.000000
-        1    1.414214
-        2    3.414214
-        3    4.828427
-        dtype: float64
-        """
-
         spark_expr = (
             F.when(
                 stf.GeometryType(self.spark.column).isin(
@@ -946,92 +883,14 @@ class GeoSeries(GeoFrame, pspd.Series):
 
     @property
     def is_valid(self) -> pspd.Series:
-        """Returns a ``Series`` of ``dtype('bool')`` with value ``True`` for
-        geometries that are valid.
-
-        Examples
-        --------
-
-        An example with one invalid polygon (a bowtie geometry crossing itself)
-        and one missing geometry:
-
-        >>> from sedona.geopandas import GeoSeries
-        >>> from shapely.geometry import Polygon
-        >>> s = GeoSeries(
-        ...     [
-        ...         Polygon([(0, 0), (1, 1), (0, 1)]),
-        ...         Polygon([(0,0), (1, 1), (1, 0), (0, 1)]),  # bowtie geometry
-        ...         Polygon([(0, 0), (2, 2), (2, 0)]),
-        ...         None
-        ...     ]
-        ... )
-        >>> s
-        0         POLYGON ((0 0, 1 1, 0 1, 0 0))
-        1    POLYGON ((0 0, 1 1, 1 0, 0 1, 0 0))
-        2         POLYGON ((0 0, 2 2, 2 0, 0 0))
-        3                                   None
-        dtype: geometry
-
-        >>> s.is_valid
-        0     True
-        1    False
-        2     True
-        3    False
-        dtype: bool
-
-        See also
-        --------
-        GeoSeries.is_valid_reason : reason for invalidity
-        """
-
         spark_col = stf.ST_IsValid(self.spark.column)
         result = self._query_geometry_column(
             spark_col,
             returns_geom=False,
         )
-        return to_bool(result)
+        return _to_bool(result)
 
     def is_valid_reason(self) -> pspd.Series:
-        """Returns a ``Series`` of strings with the reason for invalidity of
-        each geometry.
-
-        Examples
-        --------
-
-        An example with one invalid polygon (a bowtie geometry crossing itself)
-        and one missing geometry:
-
-        >>> from sedona.geopandas import GeoSeries
-        >>> from shapely.geometry import Polygon
-        >>> s = GeoSeries(
-        ...     [
-        ...         Polygon([(0, 0), (1, 1), (0, 1)]),
-        ...         Polygon([(0,0), (1, 1), (1, 0), (0, 1)]),  # bowtie geometry
-        ...         Polygon([(0, 0), (2, 2), (2, 0)]),
-        ...         Polygon([(0, 0), (2, 0), (1, 1), (2, 2), (0, 2), (1, 1), (0, 0)]),
-        ...         None
-        ...     ]
-        ... )
-        >>> s
-        0         POLYGON ((0 0, 1 1, 0 1, 0 0))
-        1    POLYGON ((0 0, 1 1, 1 0, 0 1, 0 0))
-        2         POLYGON ((0 0, 2 2, 2 0, 0 0))
-        3                                   None
-        dtype: geometry
-
-        >>> s.is_valid_reason()
-        0    Valid Geometry
-        1    Self-intersection at or near point (0.5, 0.5, NaN)
-        2    Valid Geometry
-        3    Ring Self-intersection at or near point (1.0, 1.0)
-        4    None
-        dtype: object
-
-        See also
-        --------
-        GeoSeries.is_valid : detect invalid geometries
-        GeoSeries.make_valid : fix invalid geometries
-        """
         spark_col = stf.ST_IsValidReason(self.spark.column)
         return self._query_geometry_column(
             spark_col,
@@ -1040,39 +899,12 @@ class GeoSeries(GeoFrame, pspd.Series):
 
     @property
     def is_empty(self) -> pspd.Series:
-        """
-        Returns a ``Series`` of ``dtype('bool')`` with value ``True`` for
-        empty geometries.
-
-        Examples
-        --------
-        An example of a GeoDataFrame with one empty point, one point and one missing
-        value:
-
-        >>> from sedona.geopandas import GeoSeries
-        >>> from shapely.geometry import Point
-        >>> geoseries = GeoSeries([Point(), Point(2, 1), None], crs="EPSG:4326")
-        >>> geoseries
-        0  POINT EMPTY
-        1  POINT (2 1)
-        2         None
-
-        >>> geoseries.is_empty
-        0     True
-        1    False
-        2    False
-        dtype: bool
-
-        See Also
-        --------
-        GeoSeries.isna : detect missing values
-        """
         spark_expr = stf.ST_IsEmpty(self.spark.column)
         result = self._query_geometry_column(
             spark_expr,
             returns_geom=False,
         )
-        return to_bool(result)
+        return _to_bool(result)
 
     def count_coordinates(self):
         # Implementation of the abstract method
@@ -1102,108 +934,6 @@ class GeoSeries(GeoFrame, pspd.Series):
         )
 
     def dwithin(self, other, distance, align=None):
-        """Returns a ``Series`` of ``dtype('bool')`` with value ``True`` for
-        each aligned geometry that is within a set distance from ``other``.
-
-        The operation works on a 1-to-1 row-wise manner:
-
-        Parameters
-        ----------
-        other : GeoSeries or geometric object
-            The GeoSeries (elementwise) or geometric object to test for
-            equality.
-        distance : float, np.array, pd.Series
-            Distance(s) to test if each geometry is within. A scalar distance will be
-            applied to all geometries. An array or Series will be applied elementwise.
-            If np.array or pd.Series are used then it must have same length as the
-            GeoSeries.
-        align : bool | None (default None)
-            If True, automatically aligns GeoSeries based on their indices.
-            If False, the order of elements is preserved. None defaults to True.
-
-        Returns
-        -------
-        Series (bool)
-
-        Examples
-        --------
-        >>> from sedona.geopandas import GeoSeries
-        >>> from shapely.geometry import Polygon, LineString, Point
-        >>> s = GeoSeries(
-        ...     [
-        ...         Polygon([(0, 0), (1, 1), (0, 1)]),
-        ...         LineString([(0, 0), (0, 2)]),
-        ...         LineString([(0, 0), (0, 1)]),
-        ...         Point(0, 1),
-        ...     ],
-        ...     index=range(0, 4),
-        ... )
-        >>> s2 = GeoSeries(
-        ...     [
-        ...         Polygon([(1, 0), (4, 2), (2, 2)]),
-        ...         Polygon([(2, 0), (3, 2), (2, 2)]),
-        ...         LineString([(2, 0), (2, 2)]),
-        ...         Point(1, 1),
-        ...     ],
-        ...     index=range(1, 5),
-        ... )
-
-        >>> s
-        0    POLYGON ((0 0, 1 1, 0 1, 0 0))
-        1             LINESTRING (0 0, 0 2)
-        2             LINESTRING (0 0, 0 1)
-        3                       POINT (0 1)
-        dtype: geometry
-
-        >>> s2
-        1    POLYGON ((1 0, 4 2, 2 2, 1 0))
-        2    POLYGON ((2 0, 3 2, 2 2, 2 0))
-        3             LINESTRING (2 0, 2 2)
-        4                       POINT (1 1)
-        dtype: geometry
-
-        We can check if each geometry of GeoSeries contains a single
-        geometry:
-
-        >>> point = Point(0, 1)
-        >>> s2.dwithin(point, 1.8)
-        1     True
-        2    False
-        3    False
-        4     True
-        dtype: bool
-
-        We can also check two GeoSeries against each other, row by row.
-        The GeoSeries above have different indices. We can either align both GeoSeries
-        based on index values and compare elements with the same index using
-        ``align=True`` or ignore index and compare elements based on their matching
-        order using ``align=False``:
-
-        >>> s.dwithin(s2, distance=1, align=True)
-        0    False
-        1     True
-        2    False
-        3    False
-        4    False
-        dtype: bool
-
-        >>> s.dwithin(s2, distance=1, align=False)
-        0     True
-        1    False
-        2    False
-        3     True
-        dtype: bool
-
-        Notes
-        -----
-        This method works in a row-wise manner. It does not check if an element
-        of one GeoSeries is within the set distance of *any* element of the other one.
-
-        See also
-        --------
-        GeoSeries.within
-        """
-
         if not isinstance(distance, (float, int)):
             raise NotImplementedError(
                 "Array-like distance for dwithin not implemented yet."
@@ -1222,107 +952,6 @@ class GeoSeries(GeoFrame, pspd.Series):
         )
 
     def difference(self, other, align=None) -> "GeoSeries":
-        """Returns a ``GeoSeries`` of the points in each aligned geometry that
-        are not in `other`.
-
-        The operation works on a 1-to-1 row-wise manner:
-
-        Unlike Geopandas, Sedona does not support this operation for GeometryCollections.
-
-        Parameters
-        ----------
-        other : Geoseries or geometric object
-            The Geoseries (elementwise) or geometric object to find the
-            difference to.
-        align : bool | None (default None)
-            If True, automatically aligns GeoSeries based on their indices. None defaults to True.
-            If False, the order of elements is preserved.
-
-        Returns
-        -------
-        GeoSeries
-
-        Examples
-        --------
-        >>> from sedona.geopandas import GeoSeries
-        >>> from shapely.geometry import Polygon, LineString, Point
-        >>> s = GeoSeries(
-        ...     [
-        ...         Polygon([(0, 0), (2, 2), (0, 2)]),
-        ...         Polygon([(0, 0), (2, 2), (0, 2)]),
-        ...         LineString([(0, 0), (2, 2)]),
-        ...         LineString([(2, 0), (0, 2)]),
-        ...         Point(0, 1),
-        ...     ],
-        ... )
-        >>> s2 = GeoSeries(
-        ...     [
-        ...         Polygon([(0, 0), (1, 1), (0, 1)]),
-        ...         LineString([(1, 0), (1, 3)]),
-        ...         LineString([(2, 0), (0, 2)]),
-        ...         Point(1, 1),
-        ...         Point(0, 1),
-        ...     ],
-        ...     index=range(1, 6),
-        ... )
-
-        >>> s
-        0    POLYGON ((0 0, 2 2, 0 2, 0 0))
-        1    POLYGON ((0 0, 2 2, 0 2, 0 0))
-        2             LINESTRING (0 0, 2 2)
-        3             LINESTRING (2 0, 0 2)
-        4                       POINT (0 1)
-        dtype: geometry
-
-        >>> s2
-        1    POLYGON ((0 0, 1 1, 0 1, 0 0))
-        2             LINESTRING (1 0, 1 3)
-        3             LINESTRING (2 0, 0 2)
-        4                       POINT (1 1)
-        5                       POINT (0 1)
-        dtype: geometry
-
-        We can do difference of each geometry and a single
-        shapely geometry:
-
-        >>> s.difference(Polygon([(0, 0), (1, 1), (0, 1)]))
-        0       POLYGON ((0 2, 2 2, 1 1, 0 1, 0 2))
-        1         POLYGON ((0 2, 2 2, 1 1, 0 1, 0 2))
-        2                       LINESTRING (1 1, 2 2)
-        3    MULTILINESTRING ((2 0, 1 1), (1 1, 0 2))
-        4                                 POINT EMPTY
-        dtype: geometry
-
-        We can also check two GeoSeries against each other, row by row.
-        The GeoSeries above have different indices. We can either align both GeoSeries
-        based on index values and compare elements with the same index using
-        ``align=True`` or ignore index and compare elements based on their matching
-        order using ``align=False``:
-
-        >>> s.difference(s2, align=True)
-        0                                        None
-        1         POLYGON ((0 2, 2 2, 1 1, 0 1, 0 2))
-        2    MULTILINESTRING ((0 0, 1 1), (1 1, 2 2))
-        3                            LINESTRING EMPTY
-        4                                 POINT (0 1)
-        5                                        None
-        dtype: geometry
-
-        >>> s.difference(s2, align=False)
-        0         POLYGON ((0 2, 2 2, 1 1, 0 1, 0 2))
-        1    POLYGON ((0 0, 0 2, 1 2, 2 2, 1 1, 0 0))
-        2    MULTILINESTRING ((0 0, 1 1), (1 1, 2 2))
-        3                       LINESTRING (2 0, 0 2)
-        4                                 POINT EMPTY
-        dtype: geometry
-
-        See Also
-        --------
-        GeoSeries.symmetric_difference
-        GeoSeries.union
-        GeoSeries.intersection
-        """
-
         other_series, extended = self._make_series_of_val(other)
         align = False if extended else align
 
@@ -1336,37 +965,12 @@ class GeoSeries(GeoFrame, pspd.Series):
 
     @property
     def is_simple(self) -> pspd.Series:
-        """Returns a ``Series`` of ``dtype('bool')`` with value ``True`` for
-        geometries that do not cross themselves.
-
-        This is meaningful only for `LineStrings` and `LinearRings`.
-
-        Examples
-        --------
-        >>> from sedona.geopandas import GeoSeries
-        >>> from shapely.geometry import LineString
-        >>> s = GeoSeries(
-        ...     [
-        ...         LineString([(0, 0), (1, 1), (1, -1), (0, 1)]),
-        ...         LineString([(0, 0), (1, 1), (1, -1)]),
-        ...     ]
-        ... )
-        >>> s
-        0    LINESTRING (0 0, 1 1, 1 -1, 0 1)
-        1         LINESTRING (0 0, 1 1, 1 -1)
-        dtype: geometry
-
-        >>> s.is_simple
-        0    False
-        1     True
-        dtype: bool
-        """
         spark_expr = stf.ST_IsSimple(self.spark.column)
         result = self._query_geometry_column(
             spark_expr,
             returns_geom=False,
         )
-        return to_bool(result)
+        return _to_bool(result)
 
     @property
     def is_ring(self):
@@ -1399,34 +1003,6 @@ class GeoSeries(GeoFrame, pspd.Series):
 
     @property
     def has_z(self) -> pspd.Series:
-        """Returns a ``Series`` of ``dtype('bool')`` with value ``True`` for
-        features that have a z-component.
-
-        Notes
-        -----
-        Every operation in GeoPandas is planar, i.e. the potential third
-        dimension is not taken into account.
-
-        Examples
-        --------
-        >>> from sedona.geopandas import GeoSeries
-        >>> from shapely.geometry import Point
-        >>> s = GeoSeries(
-        ...     [
-        ...         Point(0, 1),
-        ...         Point(0, 1, 2),
-        ...     ]
-        ... )
-        >>> s
-        0        POINT (0 1)
-        1    POINT Z (0 1 2)
-        dtype: geometry
-
-        >>> s.has_z
-        0    False
-        1     True
-        dtype: bool
-        """
         spark_expr = stf.ST_HasZ(self.spark.column)
         return self._query_geometry_column(
             spark_expr,
@@ -1438,79 +1014,6 @@ class GeoSeries(GeoFrame, pspd.Series):
         raise NotImplementedError("This method is not implemented yet.")
 
     def get_geometry(self, index) -> "GeoSeries":
-        """Returns the n-th geometry from a collection of geometries (0-indexed).
-
-        If the index is non-negative, it returns the geometry at that index.
-        If the index is negative, it counts backward from the end of the collection (e.g., -1 returns the last geometry).
-        Returns None if the index is out of bounds.
-
-        Note: Simple geometries act as length-1 collections
-
-        Note: Using Shapely < 2.0, may lead to different results for empty simple geometries due to how
-        shapely interprets them.
-
-        Parameters
-        ----------
-        index : int or array_like
-            Position of a geometry to be retrieved within its collection
-
-        Returns
-        -------
-        GeoSeries
-
-        Notes
-        -----
-        Simple geometries act as collections of length 1. Any out-of-range index value
-        returns None.
-
-        Examples
-        --------
-        >>> from shapely.geometry import Point, MultiPoint, GeometryCollection
-        >>> s = geopandas.GeoSeries(
-        ...     [
-        ...         Point(0, 0),
-        ...         MultiPoint([(0, 0), (1, 1), (0, 1), (1, 0)]),
-        ...         GeometryCollection(
-        ...             [MultiPoint([(0, 0), (1, 1), (0, 1), (1, 0)]), Point(0, 1)]
-        ...         ),
-        ...         Polygon(),
-        ...         GeometryCollection(),
-        ...     ]
-        ... )
-        >>> s
-        0                                          POINT (0 0)
-        1              MULTIPOINT ((0 0), (1 1), (0 1), (1 0))
-        2    GEOMETRYCOLLECTION (MULTIPOINT ((0 0), (1 1), ...
-        3                                        POLYGON EMPTY
-        4                             GEOMETRYCOLLECTION EMPTY
-        dtype: geometry
-
-        >>> s.get_geometry(0)
-        0                                POINT (0 0)
-        1                                POINT (0 0)
-        2    MULTIPOINT ((0 0), (1 1), (0 1), (1 0))
-        3                              POLYGON EMPTY
-        4                                       None
-        dtype: geometry
-
-        >>> s.get_geometry(1)
-        0           None
-        1    POINT (1 1)
-        2    POINT (0 1)
-        3           None
-        4           None
-        dtype: geometry
-
-        >>> s.get_geometry(-1)
-        0    POINT (0 0)
-        1    POINT (1 0)
-        2    POINT (0 1)
-        3  POLYGON EMPTY
-        4           None
-        dtype: geometry
-
-        """
-
         # Sedona errors on negative indexes, so we use a case statement to handle it ourselves
         spark_expr = stf.ST_GeometryN(
             F.col("L"),
@@ -1537,38 +1040,6 @@ class GeoSeries(GeoFrame, pspd.Series):
 
     @property
     def boundary(self) -> "GeoSeries":
-        """Returns a ``GeoSeries`` of lower dimensional objects representing
-        each geometry's set-theoretic `boundary`.
-
-        Examples
-        --------
-
-        >>> from sedona.geopandas import GeoSeries
-        >>> from shapely.geometry import Polygon, LineString, Point
-        >>> s = GeoSeries(
-        ...     [
-        ...         Polygon([(0, 0), (1, 1), (0, 1)]),
-        ...         LineString([(0, 0), (1, 1), (1, 0)]),
-        ...         Point(0, 0),
-        ...     ]
-        ... )
-        >>> s
-        0    POLYGON ((0 0, 1 1, 0 1, 0 0))
-        1        LINESTRING (0 0, 1 1, 1 0)
-        2                       POINT (0 0)
-        dtype: geometry
-
-        >>> s.boundary
-        0    LINESTRING (0 0, 1 1, 0 1, 0 0)
-        1          MULTIPOINT ((0 0), (1 0))
-        2           GEOMETRYCOLLECTION EMPTY
-        dtype: geometry
-
-        See also
-        --------
-        GeoSeries.exterior : outer boundary (without interior rings)
-
-        """
         # Geopandas and shapely return NULL for GeometryCollections, so we handle it separately
         # https://shapely.readthedocs.io/en/stable/reference/shapely.boundary.html
         spark_expr = F.when(
@@ -1581,39 +1052,6 @@ class GeoSeries(GeoFrame, pspd.Series):
 
     @property
     def centroid(self) -> "GeoSeries":
-        """Returns a ``GeoSeries`` of points representing the centroid of each
-        geometry.
-
-        Note that centroid does not have to be on or within original geometry.
-
-        Examples
-        --------
-
-        >>> from sedona.geopandas import GeoSeries
-        >>> from shapely.geometry import Polygon, LineString, Point
-        >>> s = GeoSeries(
-        ...     [
-        ...         Polygon([(0, 0), (1, 1), (0, 1)]),
-        ...         LineString([(0, 0), (1, 1), (1, 0)]),
-        ...         Point(0, 0),
-        ...     ]
-        ... )
-        >>> s
-        0    POLYGON ((0 0, 1 1, 0 1, 0 0))
-        1        LINESTRING (0 0, 1 1, 1 0)
-        2                       POINT (0 0)
-        dtype: geometry
-
-        >>> s.centroid
-        0    POINT (0.33333 0.66667)
-        1        POINT (0.70711 0.5)
-        2                POINT (0 0)
-        dtype: geometry
-
-        See also
-        --------
-        GeoSeries.representative_point : point guaranteed to be within each geometry
-        """
         spark_expr = stf.ST_Centroid(self.spark.column)
         return self._query_geometry_column(
             spark_expr,
@@ -1643,44 +1081,6 @@ class GeoSeries(GeoFrame, pspd.Series):
 
     @property
     def envelope(self) -> "GeoSeries":
-        """Returns a ``GeoSeries`` of geometries representing the envelope of
-        each geometry.
-
-        The envelope of a geometry is the bounding rectangle. That is, the
-        point or smallest rectangular polygon (with sides parallel to the
-        coordinate axes) that contains the geometry.
-
-        Examples
-        --------
-
-        >>> from sedona.geopandas import GeoSeries
-        >>> from shapely.geometry import Polygon, LineString, Point, MultiPoint
-        >>> s = GeoSeries(
-        ...     [
-        ...         Polygon([(0, 0), (1, 1), (0, 1)]),
-        ...         LineString([(0, 0), (1, 1), (1, 0)]),
-        ...         MultiPoint([(0, 0), (1, 1)]),
-        ...         Point(0, 0),
-        ...     ]
-        ... )
-        >>> s
-        0    POLYGON ((0 0, 1 1, 0 1, 0 0))
-        1        LINESTRING (0 0, 1 1, 1 0)
-        2         MULTIPOINT ((0 0), (1 1))
-        3                       POINT (0 0)
-        dtype: geometry
-
-        >>> s.envelope
-        0    POLYGON ((0 0, 1 0, 1 1, 0 1, 0 0))
-        1    POLYGON ((0 0, 1 0, 1 1, 0 1, 0 0))
-        2    POLYGON ((0 0, 1 0, 1 1, 0 1, 0 0))
-        3                            POINT (0 0)
-        dtype: geometry
-
-        See also
-        --------
-        GeoSeries.convex_hull : convex hull geometry
-        """
         spark_expr = stf.ST_Envelope(self.spark.column)
         return self._query_geometry_column(
             spark_expr,
@@ -1738,65 +1138,6 @@ class GeoSeries(GeoFrame, pspd.Series):
         raise NotImplementedError("This method is not implemented yet.")
 
     def make_valid(self, *, method="linework", keep_collapsed=True) -> "GeoSeries":
-        """Repairs invalid geometries.
-
-        Returns a ``GeoSeries`` with valid geometries.
-
-        If the input geometry is already valid, then it will be preserved.
-        In many cases, in order to create a valid geometry, the input
-        geometry must be split into multiple parts or multiple geometries.
-        If the geometry must be split into multiple parts of the same type
-        to be made valid, then a multi-part geometry will be returned
-        (e.g. a MultiPolygon).
-        If the geometry must be split into multiple parts of different types
-        to be made valid, then a GeometryCollection will be returned.
-
-        In Sedona, only the 'structure' method is available:
-
-        * the 'structure' algorithm tries to reason from the structure of the
-          input to find the 'correct' repair: exterior rings bound area,
-          interior holes exclude area. It first makes all rings valid, then
-          shells are merged and holes are subtracted from the shells to
-          generate valid result. It assumes that holes and shells are correctly
-          categorized in the input geometry.
-
-        Parameters
-        ----------
-        method : {'linework', 'structure'}, default 'linework'
-            Algorithm to use when repairing geometry. Sedona Geopandas only supports the 'structure' method.
-            The default method is "linework" to match compatibility with Geopandas, but it must be explicitly set to
-            'structure' to use the Sedona implementation.
-
-        keep_collapsed : bool, default True
-            For the 'structure' method, True will keep components that have
-            collapsed into a lower dimensionality. For example, a ring
-            collapsing to a line, or a line collapsing to a point.
-
-        Examples
-        --------
-
-        >>> from sedona.geopandas import GeoSeries
-        >>> from shapely.geometry import MultiPolygon, Polygon, LineString, Point
-        >>> s = GeoSeries(
-        ...     [
-        ...         Polygon([(0, 0), (0, 2), (1, 1), (2, 2), (2, 0), (1, 1), (0, 0)]),
-        ...         Polygon([(0, 2), (0, 1), (2, 0), (0, 0), (0, 2)]),
-        ...         LineString([(0, 0), (1, 1), (1, 0)]),
-        ...     ],
-        ... )
-        >>> s
-        0    POLYGON ((0 0, 0 2, 1 1, 2 2, 2 0, 1 1, 0 0))
-        1              POLYGON ((0 2, 0 1, 2 0, 0 0, 0 2))
-        2                       LINESTRING (0 0, 1 1, 1 0)
-        dtype: geometry
-
-        >>> s.make_valid()
-        0    MULTIPOLYGON (((1 1, 0 0, 0 2, 1 1)), ((2 0, 1...
-        1                       POLYGON ((0 1, 2 0, 0 0, 0 1))
-        2                           LINESTRING (0 0, 1 1, 1 0)
-        dtype: geometry
-        """
-
         if method != "structure":
             raise ValueError(
                 "Sedona only supports the 'structure' method for make_valid"
@@ -1842,34 +1183,6 @@ class GeoSeries(GeoFrame, pspd.Series):
         raise NotImplementedError("This method is not implemented yet.")
 
     def union_all(self, method="unary", grid_size=None) -> BaseGeometry:
-        """Returns a geometry containing the union of all geometries in the
-        ``GeoSeries``.
-
-        Sedona does not support the method or grid_size argument, so the user does not need to manually
-        decide the algorithm being used.
-
-        Parameters
-        ----------
-        method : str (default ``"unary"``)
-            Not supported in Sedona.
-
-        grid_size : float, default None
-            Not supported in Sedona.
-
-        Examples
-        --------
-
-        >>> from sedona.geopandas import GeoSeries
-        >>> from shapely.geometry import box
-        >>> s = GeoSeries([box(0, 0, 1, 1), box(0, 0, 2, 2)])
-        >>> s
-        0    POLYGON ((1 0, 1 1, 0 1, 0 0, 1 0))
-        1    POLYGON ((2 0, 2 2, 0 2, 0 0, 2 0))
-        dtype: geometry
-
-        >>> s.union_all()
-        <POLYGON ((0 1, 0 2, 2 2, 2 0, 1 0, 0 0, 0 1))>
-        """
         if grid_size is not None:
             raise NotImplementedError("Sedona does not support the grid_size argument")
         if method != "unary":
@@ -1894,111 +1207,6 @@ class GeoSeries(GeoFrame, pspd.Series):
         return geom
 
     def crosses(self, other, align=None) -> pspd.Series:
-        """Returns a ``Series`` of ``dtype('bool')`` with value ``True`` for
-        each aligned geometry that cross `other`.
-
-        An object is said to cross `other` if its `interior` intersects the
-        `interior` of the other but does not contain it, and the dimension of
-        the intersection is less than the dimension of the one or the other.
-
-        Note: Unlike Geopandas, Sedona's implementation always return NULL when GeometryCollection is involved.
-
-        The operation works on a 1-to-1 row-wise manner.
-
-        Parameters
-        ----------
-        other : GeoSeries or geometric object
-            The GeoSeries (elementwise) or geometric object to test if is
-            crossed.
-        align : bool | None (default None)
-            If True, automatically aligns GeoSeries based on their indices. None defaults to True.
-            If False, the order of elements is preserved.
-
-        Returns
-        -------
-        Series (bool)
-
-        Examples
-        --------
-
-        >>> from sedona.geopandas import GeoSeries
-        >>> from shapely.geometry import Polygon, LineString, Point
-        >>> s = GeoSeries(
-        ...     [
-        ...         Polygon([(0, 0), (2, 2), (0, 2)]),
-        ...         LineString([(0, 0), (2, 2)]),
-        ...         LineString([(2, 0), (0, 2)]),
-        ...         Point(0, 1),
-        ...     ],
-        ... )
-        >>> s2 = GeoSeries(
-        ...     [
-        ...         LineString([(1, 0), (1, 3)]),
-        ...         LineString([(2, 0), (0, 2)]),
-        ...         Point(1, 1),
-        ...         Point(0, 1),
-        ...     ],
-        ...     index=range(1, 5),
-        ... )
-
-        >>> s
-        0    POLYGON ((0 0, 2 2, 0 2, 0 0))
-        1             LINESTRING (0 0, 2 2)
-        2             LINESTRING (2 0, 0 2)
-        3                       POINT (0 1)
-        dtype: geometry
-        >>> s2
-        1    LINESTRING (1 0, 1 3)
-        2    LINESTRING (2 0, 0 2)
-        3              POINT (1 1)
-        4              POINT (0 1)
-        dtype: geometry
-
-        We can check if each geometry of GeoSeries crosses a single
-        geometry:
-
-        >>> line = LineString([(-1, 1), (3, 1)])
-        >>> s.crosses(line)
-        0     True
-        1     True
-        2     True
-        3    False
-        dtype: bool
-
-        We can also check two GeoSeries against each other, row by row.
-        The GeoSeries above have different indices. We can either align both GeoSeries
-        based on index values and compare elements with the same index using
-        ``align=True`` or ignore index and compare elements based on their matching
-        order using ``align=False``:
-
-        >>> s.crosses(s2, align=True)
-        0    False
-        1     True
-        2    False
-        3    False
-        4    False
-        dtype: bool
-
-        >>> s.crosses(s2, align=False)
-        0     True
-        1     True
-        2    False
-        3    False
-        dtype: bool
-
-        Notice that a line does not cross a point that it contains.
-
-        Notes
-        -----
-        This method works in a row-wise manner. It does not check if an element
-        of one GeoSeries ``crosses`` *any* element of the other one.
-
-        See also
-        --------
-        GeoSeries.disjoint
-        GeoSeries.intersects
-
-        """
         # Sedona does not support GeometryCollection (errors), so we return NULL for now to avoid error
         other_series, extended = self._make_series_of_val(other)
         align = False if extended else align
@@ -2015,7 +1223,7 @@ class GeoSeries(GeoFrame, pspd.Series):
             default_val=False,
         )
 
-        return to_bool(result)
+        return _to_bool(result)
 
     def disjoint(self, other, align=None):
         # Implementation of the abstract method
@@ -2024,108 +1232,6 @@ class GeoSeries(GeoFrame, pspd.Series):
     def intersects(
         self, other: Union["GeoSeries", BaseGeometry], align: Union[bool, None] = None
     ) -> pspd.Series:
-        """Returns a ``Series`` of ``dtype('bool')`` with value ``True`` for
-        each aligned geometry that intersects `other`.
-
-        An object is said to intersect `other` if its `boundary` and `interior`
-        intersects in any way with those of the other.
-
-        The operation works on a 1-to-1 row-wise manner.
-
-        Parameters
-        ----------
-        other : GeoSeries or geometric object
-            The GeoSeries (elementwise) or geometric object to test if is
-            intersected.
-        align : bool | None (default None)
-            If True, automatically aligns GeoSeries based on their indices. None defaults to True.
-            If False, the order of elements is preserved.
-
-        Returns
-        -------
-        Series (bool)
-
-        Examples
-        --------
-        >>> from sedona.geopandas import GeoSeries
-        >>> from shapely.geometry import Polygon, LineString, Point
-        >>> s = GeoSeries(
-        ...     [
-        ...         Polygon([(0, 0), (2, 2), (0, 2)]),
-        ...         LineString([(0, 0), (2, 2)]),
-        ...         LineString([(2, 0), (0, 2)]),
-        ...         Point(0, 1),
-        ...     ],
-        ... )
-        >>> s2 = GeoSeries(
-        ...     [
-        ...         LineString([(1, 0), (1, 3)]),
-        ...         LineString([(2, 0), (0, 2)]),
-        ...         Point(1, 1),
-        ...         Point(0, 1),
-        ...     ],
-        ...     index=range(1, 5),
-        ... )
-
-        >>> s
-        0    POLYGON ((0 0, 2 2, 0 2, 0 0))
-        1             LINESTRING (0 0, 2 2)
-        2             LINESTRING (2 0, 0 2)
-        3                       POINT (0 1)
-        dtype: geometry
-
-        >>> s2
-        1    LINESTRING (1 0, 1 3)
-        2    LINESTRING (2 0, 0 2)
-        3              POINT (1 1)
-        4              POINT (0 1)
-        dtype: geometry
-
-        We can check if each geometry of GeoSeries crosses a single
-        geometry:
-
-        >>> line = LineString([(-1, 1), (3, 1)])
-        >>> s.intersects(line)
-        0    True
-        1    True
-        2    True
-        3    True
-        dtype: bool
-
-        We can also check two GeoSeries against each other, row by row.
-        The GeoSeries above have different indices. We can either align both GeoSeries
-        based on index values and compare elements with the same index using
-        ``align=True`` or ignore index and compare elements based on their matching
-        order using ``align=False``:
-
-        >>> s.intersects(s2, align=True)
-        0    False
-        1     True
-        2     True
-        3    False
-        4    False
-        dtype: bool
-
-        >>> s.intersects(s2, align=False)
-        0    True
-        1    True
-        2    True
-        3    True
-        dtype: bool
-
-        Notes
-        -----
-        This method works in a row-wise manner. It does not check if an element
-        of one GeoSeries ``crosses`` *any* element of the other one.
-
-        See also
-        --------
-        GeoSeries.disjoint
-        GeoSeries.crosses
-        GeoSeries.touches
-        GeoSeries.intersection
-        """
-
         other_series, extended = self._make_series_of_val(other)
         align = False if extended else align
 
@@ -2136,97 +1242,9 @@ class GeoSeries(GeoFrame, pspd.Series):
             align,
             default_val=False,
         )
-        return to_bool(result)
+        return _to_bool(result)
 
     def overlaps(self, other, align=None) -> pspd.Series:
-        """Returns True for all aligned geometries that overlap other, else False.
-
-        In the original Geopandas, Geometries overlap if they have more than one but not all
-        points in common, have the same dimension, and the intersection of the
-        interiors of the geometries has the same dimension as the geometries
-        themselves.
-
-        However, in Sedona, we return True in the case where the geometries points match.
-
-        Note: Sedona's behavior may also differ from Geopandas for GeometryCollections.
-
-        The operation works on a 1-to-1 row-wise manner.
-
-        Parameters
-        ----------
-        other : GeoSeries or geometric object
-            The GeoSeries (elementwise) or geometric object to test if
-            overlaps.
-        align : bool | None (default None)
-            If True, automatically aligns GeoSeries based on their indices. None defaults to True.
-            If False, the order of elements is preserved.
-
-        Returns
-        -------
-        Series (bool)
-
-        Examples
-        --------
-        >>> from sedona.geopandas import GeoSeries
-        >>> from shapely.geometry import Polygon, LineString, MultiPoint, Point
-        >>> s = GeoSeries(
-        ...     [
-        ...         Polygon([(0, 0), (2, 2), (0, 2)]),
-        ...         Polygon([(0, 0), (2, 2), (0, 2)]),
-        ...         LineString([(0, 0), (2, 2)]),
-        ...         MultiPoint([(0, 0), (0, 1)]),
-        ...     ],
-        ... )
-        >>> s2 = GeoSeries(
-        ...     [
-        ...         Polygon([(0, 0), (2, 0), (0, 2)]),
-        ...         LineString([(0, 1), (1, 1)]),
-        ...         LineString([(1, 1), (3, 3)]),
-        ...         Point(0, 1),
-        ...     ],
-        ... )
-
-        We can check if each geometry of GeoSeries overlaps a single
-        geometry:
-
-        >>> polygon = Polygon([(0, 0), (1, 0), (1, 1), (0, 1)])
-        >>> s.overlaps(polygon)
-        0     True
-        1     True
-        2    False
-        3    False
-        dtype: bool
-
-        We can also check two GeoSeries against each other, row by row.
-        The GeoSeries above have different indices. We align both GeoSeries
-        based on index values and compare elements with the same index.
-
-        >>> s.overlaps(s2)
-        0    False
-        1     True
-        2    False
-        3    False
-        4    False
-        dtype: bool
-
-        >>> s.overlaps(s2, align=False)
-        0     True
-        1    False
-        2     True
-        3    False
-        dtype: bool
-
-        Notes
-        -----
-        This method works in a row-wise manner. It does not check if an element
-        of one GeoSeries ``overlaps`` *any* element of the other one.
-
-        See also
-        --------
-        GeoSeries.crosses
-        GeoSeries.intersects
-
-        """
         # Note: We cannot efficiently match geopandas behavior because Sedona's ST_Overlaps returns True for equal geometries
         # ST_Overlaps(`L`, `R`) AND ST_Equals(`L`, `R`) does not work because ST_Equals errors on invalid geometries
 
@@ -2240,112 +1258,9 @@ class GeoSeries(GeoFrame, pspd.Series):
             align,
             default_val=False,
         )
-        return to_bool(result)
+        return _to_bool(result)
 
     def touches(self, other, align=None) -> pspd.Series:
-        """Returns a ``Series`` of ``dtype('bool')`` with value ``True`` for
-        each aligned geometry that touches `other`.
-
-        An object is said to touch `other` if it has at least one point in
-        common with `other` and its interior does not intersect with any part
-        of the other. Overlapping features therefore do not touch.
-
-        Note: Sedona's behavior may also differ from Geopandas for GeometryCollections.
-
-        The operation works on a 1-to-1 row-wise manner.
-
-        Parameters
-        ----------
-        other : GeoSeries or geometric object
-            The GeoSeries (elementwise) or geometric object to test if is
-            touched.
-        align : bool | None (default None)
-            If True, automatically aligns GeoSeries based on their indices. None defaults to True.
-            If False, the order of elements is preserved.
-
-        Returns
-        -------
-        Series (bool)
-
-        Examples
-        --------
-        >>> from shapely.geometry import Polygon, LineString, MultiPoint, Point
-        >>> s = GeoSeries(
-        ...     [
-        ...         Polygon([(0, 0), (2, 2), (0, 2)]),
-        ...         Polygon([(0, 0), (2, 2), (0, 2)]),
-        ...         LineString([(0, 0), (2, 2)]),
-        ...         MultiPoint([(0, 0), (0, 1)]),
-        ...     ],
-        ... )
-        >>> s2 = GeoSeries(
-        ...     [
-        ...         Polygon([(0, 0), (-2, 0), (0, -2)]),
-        ...         LineString([(0, 1), (1, 1)]),
-        ...         LineString([(1, 1), (3, 0)]),
-        ...         Point(0, 1),
-        ...     ],
-        ...     index=range(1, 5),
-        ... )
-
-        >>> s
-        0    POLYGON ((0 0, 2 2, 0 2, 0 0))
-        1    POLYGON ((0 0, 2 2, 0 2, 0 0))
-        2             LINESTRING (0 0, 2 2)
-        3         MULTIPOINT ((0 0), (0 1))
-        dtype: geometry
-
-        >>> s2
-        1    POLYGON ((0 0, -2 0, 0 -2, 0 0))
-        2               LINESTRING (0 1, 1 1)
-        3               LINESTRING (1 1, 3 0)
-        4                         POINT (0 1)
-        dtype: geometry
-
-        We can check if each geometry of GeoSeries touches a single
-        geometry:
-
-        >>> line = LineString([(0, 0), (-1, -2)])
-        >>> s.touches(line)
-        0    True
-        1    True
-        2    True
-        3    True
-        dtype: bool
-
-        We can also check two GeoSeries against each other, row by row.
-        The GeoSeries above have different indices. We can either align both GeoSeries
-        based on index values and compare elements with the same index using
-        ``align=True`` or ignore index and compare elements based on their matching
-        order using ``align=False``:
-
-        >>> s.touches(s2, align=True)
-        0    False
-        1     True
-        2     True
-        3    False
-        4    False
-        dtype: bool
-
-        >>> s.touches(s2, align=False)
-        0     True
-        1    False
-        2     True
-        3    False
-        dtype: bool
-
-        Notes
-        -----
-        This method works in a row-wise manner. It does not check if an element
-        of one GeoSeries ``touches`` *any* element of the other one.
-
-        See also
-        --------
-        GeoSeries.overlaps
-        GeoSeries.intersects
-
-        """
-
         other_series, extended = self._make_series_of_val(other)
         align = False if extended else align
 
@@ -2356,115 +1271,9 @@ class GeoSeries(GeoFrame, pspd.Series):
             align,
             default_val=False,
         )
-        return to_bool(result)
+        return _to_bool(result)
 
     def within(self, other, align=None) -> pspd.Series:
-        """Returns a ``Series`` of ``dtype('bool')`` with value ``True`` for
-        each aligned geometry that is within `other`.
-
-        An object is said to be within `other` if at least one of its points is located
-        in the `interior` and no points are located in the `exterior` of the other.
-        If either object is empty, this operation returns ``False``.
-
-        This is the inverse of `contains` in the sense that the
-        expression ``a.within(b) == b.contains(a)`` always evaluates to
-        ``True``.
-
-        Note: Sedona's behavior may also differ from Geopandas for GeometryCollections and for geometries that are equal.
-
-        The operation works on a 1-to-1 row-wise manner.
-
-        Parameters
-        ----------
-        other : GeoSeries or geometric object
-            The GeoSeries (elementwise) or geometric object to test if each
-            geometry is within.
-        align : bool | None (default None)
-            If True, automatically aligns GeoSeries based on their indices. None defaults to True.
-            If False, the order of elements is preserved.
-
-        Returns
-        -------
-        Series (bool)
-
-
-        Examples
-        --------
-        >>> from shapely.geometry import Polygon, LineString, Point
-        >>> s = GeoSeries(
-        ...     [
-        ...         Polygon([(0, 0), (2, 2), (0, 2)]),
-        ...         Polygon([(0, 0), (1, 2), (0, 2)]),
-        ...         LineString([(0, 0), (0, 2)]),
-        ...         Point(0, 1),
-        ...     ],
-        ... )
-        >>> s2 = GeoSeries(
-        ...     [
-        ...         Polygon([(0, 0), (1, 1), (0, 1)]),
-        ...         LineString([(0, 0), (0, 2)]),
-        ...         LineString([(0, 0), (0, 1)]),
-        ...         Point(0, 1),
-        ...     ],
-        ...     index=range(1, 5),
-        ... )
-
-        >>> s
-        0    POLYGON ((0 0, 2 2, 0 2, 0 0))
-        1    POLYGON ((0 0, 1 2, 0 2, 0 0))
-        2             LINESTRING (0 0, 0 2)
-        3                       POINT (0 1)
-        dtype: geometry
-
-        >>> s2
-        1    POLYGON ((0 0, 1 1, 0 1, 0 0))
-        2             LINESTRING (0 0, 0 2)
-        3             LINESTRING (0 0, 0 1)
-        4                       POINT (0 1)
-        dtype: geometry
-
-        We can check if each geometry of GeoSeries is within a single
-        geometry:
-
-        >>> polygon = Polygon([(0, 0), (2, 2), (0, 2)])
-        >>> s.within(polygon)
-        0     True
-        1     True
-        2    False
-        3    False
-        dtype: bool
-
-        We can also check two GeoSeries against each other, row by row.
-        The GeoSeries above have different indices. We can either align both GeoSeries
-        based on index values and compare elements with the same index using
-        ``align=True`` or ignore index and compare elements based on their matching
-        order using ``align=False``:
-
-        >>> s2.within(s)
-        0    False
-        1    False
-        2     True
-        3    False
-        4    False
-        dtype: bool
-
-        >>> s2.within(s, align=False)
-        1     True
-        2    False
-        3     True
-        4     True
-        dtype: bool
-
-        Notes
-        -----
-        This method works in a row-wise manner. It does not check if an element
-        of one GeoSeries is ``within`` any element of the other one.
-
-        See also
-        --------
-        GeoSeries.contains
-        """
-
         other_series, extended = self._make_series_of_val(other)
         align = False if extended else align
 
@@ -2475,116 +1284,9 @@ class GeoSeries(GeoFrame, pspd.Series):
             align,
             default_val=False,
         )
-        return to_bool(result)
+        return _to_bool(result)
 
     def covers(self, other, align=None) -> pspd.Series:
-        """
-        Returns a ``Series`` of ``dtype('bool')`` with value ``True`` for
-        each aligned geometry that is entirely covering `other`.
-
-        An object A is said to cover another object B if no points of B lie
-        in the exterior of A.
-        If either object is empty, this operation returns ``False``.
-
-        Note: Sedona's implementation instead returns False for identical geometries.
-        Sedona's behavior may also differ from Geopandas for GeometryCollections.
-
-        The operation works on a 1-to-1 row-wise manner.
-
-        See
-        https://lin-ear-th-inking.blogspot.com/2007/06/subtleties-of-ogc-covers-spatial.html
-        for reference.
-
-        Parameters
-        ----------
-        other : Geoseries or geometric object
-            The Geoseries (elementwise) or geometric object to check is being covered.
-        align : bool | None (default None)
-            If True, automatically aligns GeoSeries based on their indices. None defaults to True.
-            If False, the order of elements is preserved.
-
-        Returns
-        -------
-        Series (bool)
-
-        Examples
-        --------
-        >>> from shapely.geometry import Polygon, LineString, Point
-        >>> s = GeoSeries(
-        ...     [
-        ...         Polygon([(0, 0), (2, 0), (2, 2), (0, 2)]),
-        ...         Polygon([(0, 0), (2, 2), (0, 2)]),
-        ...         LineString([(0, 0), (2, 2)]),
-        ...         Point(0, 0),
-        ...     ],
-        ... )
-        >>> s2 = GeoSeries(
-        ...     [
-        ...         Polygon([(0.5, 0.5), (1.5, 0.5), (1.5, 1.5), (0.5, 1.5)]),
-        ...         Polygon([(0, 0), (2, 0), (2, 2), (0, 2)]),
-        ...         LineString([(1, 1), (1.5, 1.5)]),
-        ...         Point(0, 0),
-        ...     ],
-        ...     index=range(1, 5),
-        ... )
-
-        >>> s
-        0    POLYGON ((0 0, 2 0, 2 2, 0 2, 0 0))
-        1         POLYGON ((0 0, 2 2, 0 2, 0 0))
-        2                  LINESTRING (0 0, 2 2)
-        3                            POINT (0 0)
-        dtype: geometry
-
-        >>> s2
-        1    POLYGON ((0.5 0.5, 1.5 0.5, 1.5 1.5, 0.5 1.5, ...
-        2                  POLYGON ((0 0, 2 0, 2 2, 0 2, 0 0))
-        3                            LINESTRING (1 1, 1.5 1.5)
-        4                                          POINT (0 0)
-        dtype: geometry
-
-        We can check if each geometry of GeoSeries covers a single
-        geometry:
-
-        >>> poly = Polygon([(0, 0), (2, 0), (2, 2), (0, 2)])
-        >>> s.covers(poly)
-        0     True
-        1    False
-        2    False
-        3    False
-        dtype: bool
-
-        We can also check two GeoSeries against each other, row by row.
-        The GeoSeries above have different indices. We can either align both GeoSeries
-        based on index values and compare elements with the same index using
-        ``align=True`` or ignore index and compare elements based on their matching
-        order using ``align=False``:
-
-        >>> s.covers(s2, align=True)
-        0    False
-        1    False
-        2    False
-        3    False
-        4    False
-        dtype: bool
-
-        >>> s.covers(s2, align=False)
-        0     True
-        1    False
-        2     True
-        3     True
-        dtype: bool
-
-        Notes
-        -----
-        This method works in a row-wise manner. It does not check if an element
-        of one GeoSeries ``covers`` any element of the other one.
-
-        See also
-        --------
-        GeoSeries.covered_by
-        GeoSeries.overlaps
-        """
-
         other_series, extended = self._make_series_of_val(other)
         align = False if extended else align
 
@@ -2595,116 +1297,9 @@ class GeoSeries(GeoFrame, pspd.Series):
             align,
             default_val=False,
         )
-        return to_bool(result)
+        return _to_bool(result)
 
     def covered_by(self, other, align=None) -> pspd.Series:
-        """
-        Returns a ``Series`` of ``dtype('bool')`` with value ``True`` for
-        each aligned geometry that is entirely covered by `other`.
-
-        An object A is said to cover another object B if no points of B lie
-        in the exterior of A.
-
-        Note: Sedona's implementation instead returns False for identical geometries.
-        Sedona's behavior may differ from Geopandas for GeometryCollections.
-
-        The operation works on a 1-to-1 row-wise manner.
-
-        See
-        https://lin-ear-th-inking.blogspot.com/2007/06/subtleties-of-ogc-covers-spatial.html
-        for reference.
-
-        Parameters
-        ----------
-        other : Geoseries or geometric object
-            The Geoseries (elementwise) or geometric object to check is being covered.
-        align : bool | None (default None)
-            If True, automatically aligns GeoSeries based on their indices. None defaults to True.
-            If False, the order of elements is preserved.
-
-        Returns
-        -------
-        Series (bool)
-
-        Examples
-        --------
-        >>> from shapely.geometry import Polygon, LineString, Point
-        >>> s = GeoSeries(
-        ...     [
-        ...         Polygon([(0.5, 0.5), (1.5, 0.5), (1.5, 1.5), (0.5, 1.5)]),
-        ...         Polygon([(0, 0), (2, 0), (2, 2), (0, 2)]),
-        ...         LineString([(1, 1), (1.5, 1.5)]),
-        ...         Point(0, 0),
-        ...     ],
-        ... )
-        >>> s2 = GeoSeries(
-        ...     [
-        ...         Polygon([(0, 0), (2, 0), (2, 2), (0, 2)]),
-        ...         Polygon([(0, 0), (2, 2), (0, 2)]),
-        ...         LineString([(0, 0), (2, 2)]),
-        ...         Point(0, 0),
-        ...     ],
-        ...     index=range(1, 5),
-        ... )
-
-        >>> s
-        0    POLYGON ((0.5 0.5, 1.5 0.5, 1.5 1.5, 0.5 1.5, ...
-        1                  POLYGON ((0 0, 2 0, 2 2, 0 2, 0 0))
-        2                            LINESTRING (1 1, 1.5 1.5)
-        3                                          POINT (0 0)
-        dtype: geometry
-        >>>
-
-        >>> s2
-        1    POLYGON ((0 0, 2 0, 2 2, 0 2, 0 0))
-        2         POLYGON ((0 0, 2 2, 0 2, 0 0))
-        3                  LINESTRING (0 0, 2 2)
-        4                            POINT (0 0)
-        dtype: geometry
-
-        We can check if each geometry of GeoSeries is covered by a single
-        geometry:
-
-        >>> poly = Polygon([(0, 0), (2, 0), (2, 2), (0, 2)])
-        >>> s.covered_by(poly)
-        0    True
-        1    True
-        2    True
-        3    True
-        dtype: bool
-
-        We can also check two GeoSeries against each other, row by row.
-        The GeoSeries above have different indices. We can either align both GeoSeries
-        based on index values and compare elements with the same index using
-        ``align=True`` or ignore index and compare elements based on their matching
-        order using ``align=False``:
-
-        >>> s.covered_by(s2, align=True)
-        0    False
-        1     True
-        2     True
-        3     True
-        4    False
-        dtype: bool
-
-        >>> s.covered_by(s2, align=False)
-        0     True
-        1    False
-        2     True
-        3     True
-        dtype: bool
-
-        Notes
-        -----
-        This method works in a row-wise manner. It does not check if an element
-        of one GeoSeries is ``covered_by`` any element of the other one.
-
-        See also
-        --------
-        GeoSeries.covers
-        GeoSeries.overlaps
-        """
-
         other_series, extended = self._make_series_of_val(other)
         align = False if extended else align
 
@@ -2715,95 +1310,9 @@ class GeoSeries(GeoFrame, pspd.Series):
             align,
             default_val=False,
         )
-        return to_bool(result)
+        return _to_bool(result)
 
     def distance(self, other, align=None) -> pspd.Series:
-        """Returns a ``Series`` containing the distance to aligned `other`.
-
-        The operation works on a 1-to-1 row-wise manner:
-
-        Parameters
-        ----------
-        other : Geoseries or geometric object
-            The Geoseries (elementwise) or geometric object to find the
-            distance to.
-        align : bool | None (default None)
-            If True, automatically aligns GeoSeries based on their indices. None defaults to True.
-            If False, the order of elements is preserved.
-
-        Returns
-        -------
-        Series (float)
-
-        Examples
-        --------
-        >>> from sedona.geopandas import GeoSeries
-        >>> from shapely.geometry import Polygon, LineString, Point
-        >>> s = GeoSeries(
-        ...     [
-        ...         Polygon([(0, 0), (1, 0), (1, 1)]),
-        ...         Polygon([(0, 0), (-1, 0), (-1, 1)]),
-        ...         LineString([(1, 1), (0, 0)]),
-        ...         Point(0, 0),
-        ...     ],
-        ... )
-        >>> s2 = GeoSeries(
-        ...     [
-        ...         Polygon([(0.5, 0.5), (1.5, 0.5), (1.5, 1.5), (0.5, 1.5)]),
-        ...         Point(3, 1),
-        ...         LineString([(1, 0), (2, 0)]),
-        ...         Point(0, 1),
-        ...     ],
-        ...     index=range(1, 5),
-        ... )
-
-        >>> s
-        0      POLYGON ((0 0, 1 0, 1 1, 0 0))
-        1    POLYGON ((0 0, -1 0, -1 1, 0 0))
-        2               LINESTRING (1 1, 0 0)
-        3                         POINT (0 0)
-        dtype: geometry
-
-        >>> s2
-        1    POLYGON ((0.5 0.5, 1.5 0.5, 1.5 1.5, 0.5 1.5, ...
-        2                                          POINT (3 1)
-        3                                LINESTRING (1 0, 2 0)
-        4                                          POINT (0 1)
-        dtype: geometry
-
-        We can check the distance of each geometry of GeoSeries to a single
-        geometry:
-
-        >>> point = Point(-1, 0)
-        >>> s.distance(point)
-        0    1.0
-        1    0.0
-        2    1.0
-        3    1.0
-        dtype: float64
-
-        We can also check two GeoSeries against each other, row by row.
-        The GeoSeries above have different indices. We can either align both GeoSeries
-        based on index values and use elements with the same index using
-        ``align=True`` or ignore index and use elements based on their matching
-        order using ``align=False``:
-
-        >>> s.distance(s2, align=True)
-        0         NaN
-        1    0.707107
-        2    2.000000
-        3    1.000000
-        4         NaN
-        dtype: float64
-
-        >>> s.distance(s2, align=False)
-        0    0.000000
-        1    3.162278
-        2    0.707107
-        3    1.000000
-        dtype: float64
-        """
-
         other_series, extended = self._make_series_of_val(other)
         align = False if extended else align
 
@@ -2819,106 +1328,6 @@ class GeoSeries(GeoFrame, pspd.Series):
     def intersection(
         self, other: Union["GeoSeries", BaseGeometry], align: Union[bool, None] = None
     ) -> "GeoSeries":
-        """Returns a ``GeoSeries`` of the intersection of points in each
-        aligned geometry with `other`.
-
-        The operation works on a 1-to-1 row-wise manner.
-
-        Parameters
-        ----------
-        other : Geoseries or geometric object
-            The Geoseries (elementwise) or geometric object to find the
-            intersection with.
-        align : bool | None (default None)
-            If True, automatically aligns GeoSeries based on their indices. None defaults to True.
-            If False, the order of elements is preserved.
-
-        Returns
-        -------
-        GeoSeries
-
-        Examples
-        --------
-        >>> from sedona.geopandas import GeoSeries
-        >>> from shapely.geometry import Polygon, LineString, Point
-        >>> s = GeoSeries(
-        ...     [
-        ...         Polygon([(0, 0), (2, 2), (0, 2)]),
-        ...         Polygon([(0, 0), (2, 2), (0, 2)]),
-        ...         LineString([(0, 0), (2, 2)]),
-        ...         LineString([(2, 0), (0, 2)]),
-        ...         Point(0, 1),
-        ...     ],
-        ... )
-        >>> s2 = GeoSeries(
-        ...     [
-        ...         Polygon([(0, 0), (1, 1), (0, 1)]),
-        ...         LineString([(1, 0), (1, 3)]),
-        ...         LineString([(2, 0), (0, 2)]),
-        ...         Point(1, 1),
-        ...         Point(0, 1),
-        ...     ],
-        ...     index=range(1, 6),
-        ... )
-
-        >>> s
-        0    POLYGON ((0 0, 2 2, 0 2, 0 0))
-        1    POLYGON ((0 0, 2 2, 0 2, 0 0))
-        2             LINESTRING (0 0, 2 2)
-        3             LINESTRING (2 0, 0 2)
-        4                       POINT (0 1)
-        dtype: geometry
-
-        >>> s2
-        1    POLYGON ((0 0, 1 1, 0 1, 0 0))
-        2             LINESTRING (1 0, 1 3)
-        3             LINESTRING (2 0, 0 2)
-        4                       POINT (1 1)
-        5                       POINT (0 1)
-        dtype: geometry
-
-        We can also do intersection of each geometry and a single
-        shapely geometry:
-
-        >>> s.intersection(Polygon([(0, 0), (1, 1), (0, 1)]))
-        0    POLYGON ((0 0, 0 1, 1 1, 0 0))
-        1    POLYGON ((0 0, 0 1, 1 1, 0 0))
-        2             LINESTRING (0 0, 1 1)
-        3                       POINT (1 1)
-        4                       POINT (0 1)
-        dtype: geometry
-
-        We can also check two GeoSeries against each other, row by row.
-        The GeoSeries above have different indices. We can either align both GeoSeries
-        based on index values and compare elements with the same index using
-        ``align=True`` or ignore index and compare elements based on their matching
-        order using ``align=False``:
-
-        >>> s.intersection(s2, align=True)
-        0                              None
-        1    POLYGON ((0 0, 0 1, 1 1, 0 0))
-        2                       POINT (1 1)
-        3             LINESTRING (2 0, 0 2)
-        4                       POINT EMPTY
-        5                              None
-        dtype: geometry
-
-        >>> s.intersection(s2, align=False)
-        0    POLYGON ((0 0, 0 1, 1 1, 0 0))
-        1             LINESTRING (1 1, 1 2)
-        2                       POINT (1 1)
-        3                       POINT (1 1)
-        4                       POINT (0 1)
-        dtype: geometry
-
-
-        See Also
-        --------
-        GeoSeries.difference
-        GeoSeries.symmetric_difference
-        GeoSeries.union
-        """
-
         other_series, extended = self._make_series_of_val(other)
         align = False if extended else align
 
@@ -2933,100 +1342,6 @@ class GeoSeries(GeoFrame, pspd.Series):
         return result
 
     def snap(self, other, tolerance, align=None) -> "GeoSeries":
-        """Snap the vertices and segments of the geometry to vertices of the reference.
-
-        Vertices and segments of the input geometry are snapped to vertices of the
-        reference geometry, returning a new geometry; the input geometries are not
-        modified. The result geometry is the input geometry with the vertices and
-        segments snapped. If no snapping occurs then the input geometry is returned
-        unchanged. The tolerance is used to control where snapping is performed.
-
-        Where possible, this operation tries to avoid creating invalid geometries;
-        however, it does not guarantee that output geometries will be valid. It is
-        the responsibility of the caller to check for and handle invalid geometries.
-
-        Because too much snapping can result in invalid geometries being created,
-        heuristics are used to determine the number and location of snapped
-        vertices that are likely safe to snap. These heuristics may omit
-        some potential snaps that are otherwise within the tolerance.
-
-        Note: Sedona's result may differ slightly from geopandas's snap() result
-        because of small differences between the underlying engines being used.
-
-        The operation works in a 1-to-1 row-wise manner:
-
-        Parameters
-        ----------
-        other : GeoSeries or geometric object
-            The Geoseries (elementwise) or geometric object to snap to.
-        tolerance : float or array like
-            Maximum distance between vertices that shall be snapped
-        align : bool | None (default None)
-            If True, automatically aligns GeoSeries based on their indices. None defaults to True.
-            If False, the order of elements is preserved.
-
-        Returns
-        -------
-        GeoSeries
-
-        Examples
-        --------
-        >>> from sedona.geopandas import GeoSeries
-        >>> from shapely import Polygon, LineString, Point
-        >>> s = GeoSeries(
-        ...     [
-        ...         Point(0.5, 2.5),
-        ...         LineString([(0.1, 0.1), (0.49, 0.51), (1.01, 0.89)]),
-        ...         Polygon([(0, 0), (0, 10), (10, 10), (10, 0), (0, 0)]),
-        ...     ],
-        ... )
-        >>> s
-        0                               POINT (0.5 2.5)
-        1    LINESTRING (0.1 0.1, 0.49 0.51, 1.01 0.89)
-        2       POLYGON ((0 0, 0 10, 10 10, 10 0, 0 0))
-        dtype: geometry
-
-        >>> s2 = GeoSeries(
-        ...     [
-        ...         Point(0, 2),
-        ...         LineString([(0, 0), (0.5, 0.5), (1.0, 1.0)]),
-        ...         Point(8, 10),
-        ...     ],
-        ...     index=range(1, 4),
-        ... )
-        >>> s2
-        1                       POINT (0 2)
-        2    LINESTRING (0 0, 0.5 0.5, 1 1)
-        3                      POINT (8 10)
-        dtype: geometry
-
-        We can snap each geometry to a single shapely geometry:
-
-        >>> s.snap(Point(0, 2), tolerance=1)
-        0                                     POINT (0 2)
-        1      LINESTRING (0.1 0.1, 0.49 0.51, 1.01 0.89)
-        2    POLYGON ((0 0, 0 2, 0 10, 10 10, 10 0, 0 0))
-        dtype: geometry
-
-        We can also snap two GeoSeries to each other, row by row.
-        The GeoSeries above have different indices. We can either align both GeoSeries
-        based on index values and snap elements with the same index using
-        ``align=True`` or ignore index and snap elements based on their matching
-        order using ``align=False``:
-
-        >>> s.snap(s2, tolerance=1, align=True)
-        0                                                 None
-        1           LINESTRING (0.1 0.1, 0.49 0.51, 1.01 0.89)
-        2    POLYGON ((0.5 0.5, 1 1, 0 10, 10 10, 10 0, 0.5...
-        3                                                 None
-        dtype: geometry
-
-        >>> s.snap(s2, tolerance=1, align=False)
-        0                                      POINT (0 2)
-        1                   LINESTRING (0 0, 0.5 0.5, 1 1)
-        2    POLYGON ((0 0, 0 10, 8 10, 10 10, 10 0, 0 0))
-        dtype: geometry
-        """
         if not isinstance(tolerance, (float, int)):
             raise NotImplementedError(
                 "Array-like values for tolerance are not supported yet."
@@ -3055,6 +1370,7 @@ class GeoSeries(GeoFrame, pspd.Series):
         align: Union[bool, None],
         returns_geom: bool = False,
         default_val: Any = None,
+        keep_name: bool = False,
     ):
         """
         Helper function to perform a row-wise operation on two GeoSeries.
@@ -3105,18 +1421,12 @@ class GeoSeries(GeoFrame, pspd.Series):
                 F.col("L").isNull() | F.col("R").isNull(),
                 default_val,
             ).otherwise(spark_col)
-            # The above is equivalent to the following:
-            f"""
-                CASE
-                    WHEN `L` IS NULL OR `R` IS NULL THEN {default_val}
-                    ELSE {spark_col}
-                END
-            """
 
         return self._query_geometry_column(
             spark_col,
             joined_df,
             returns_geom=returns_geom,
+            keep_name=keep_name,
         )
 
     def intersection_all(self):
@@ -3128,116 +1438,6 @@ class GeoSeries(GeoFrame, pspd.Series):
     # ============================================================================
 
     def contains(self, other, align=None) -> pspd.Series:
-        """Returns a ``Series`` of ``dtype('bool')`` with value ``True`` for
-        each aligned geometry that contains `other`.
-
-        An object is said to contain `other` if at least one point of `other` lies in
-        the interior and no points of `other` lie in the exterior of the object.
-        (Therefore, any given polygon does not contain its own boundary - there is not
-        any point that lies in the interior.)
-        If either object is empty, this operation returns ``False``.
-
-        This is the inverse of `within` in the sense that the expression
-        ``a.contains(b) == b.within(a)`` always evaluates to ``True``.
-
-        Note: Sedona's implementation instead returns False for identical geometries.
-
-        The operation works on a 1-to-1 row-wise manner.
-
-        Parameters
-        ----------
-        other : GeoSeries or geometric object
-            The GeoSeries (elementwise) or geometric object to test if it
-            is contained.
-        align : bool | None (default None)
-            If True, automatically aligns GeoSeries based on their indices. None defaults to True.
-            If False, the order of elements is preserved.
-
-        Returns
-        -------
-        Series (bool)
-
-        Examples
-        --------
-
-        >>> from sedona.geopandas import GeoSeries
-        >>> from shapely.geometry import Polygon, LineString, Point
-        >>> s = GeoSeries(
-        ...     [
-        ...         Polygon([(0, 0), (1, 1), (0, 1)]),
-        ...         LineString([(0, 0), (0, 2)]),
-        ...         LineString([(0, 0), (0, 1)]),
-        ...         Point(0, 1),
-        ...     ],
-        ...     index=range(0, 4),
-        ... )
-        >>> s2 = GeoSeries(
-        ...     [
-        ...         Polygon([(0, 0), (2, 2), (0, 2)]),
-        ...         Polygon([(0, 0), (1, 2), (0, 2)]),
-        ...         LineString([(0, 0), (0, 2)]),
-        ...         Point(0, 1),
-        ...     ],
-        ...     index=range(1, 5),
-        ... )
-
-        >>> s
-        0    POLYGON ((0 0, 1 1, 0 1, 0 0))
-        1             LINESTRING (0 0, 0 2)
-        2             LINESTRING (0 0, 0 1)
-        3                       POINT (0 1)
-        dtype: geometry
-
-        >>> s2
-        1    POLYGON ((0 0, 2 2, 0 2, 0 0))
-        2    POLYGON ((0 0, 1 2, 0 2, 0 0))
-        3             LINESTRING (0 0, 0 2)
-        4                       POINT (0 1)
-        dtype: geometry
-
-        We can check if each geometry of GeoSeries contains a single
-        geometry:
-
-        >>> point = Point(0, 1)
-        >>> s.contains(point)
-        0    False
-        1     True
-        2    False
-        3     True
-        dtype: bool
-
-        We can also check two GeoSeries against each other, row by row.
-        The GeoSeries above have different indices. We can either align both GeoSeries
-        based on index values and compare elements with the same index using
-        ``align=True`` or ignore index and compare elements based on their matching
-        order using ``align=False``:
-
-        >>> s2.contains(s, align=True)
-        0    False
-        1    False
-        2    False
-        3     True
-        4    False
-        dtype: bool
-
-        >>> s2.contains(s, align=False)
-        1     True
-        2    False
-        3     True
-        4     True
-        dtype: bool
-
-        Notes
-        -----
-        This method works in a row-wise manner. It does not check if an element
-        of one GeoSeries ``contains`` any element of the other one.
-
-        See also
-        --------
-        GeoSeries.contains_properly
-        GeoSeries.within
-        """
-
         other, extended = self._make_series_of_val(other)
         align = False if extended else align
 
@@ -3249,7 +1449,7 @@ class GeoSeries(GeoFrame, pspd.Series):
             returns_geom=False,
             default_val=False,
         )
-        return to_bool(result)
+        return _to_bool(result)
 
     def contains_properly(self, other, align=None):
         # Implementation of the abstract method
@@ -3270,29 +1470,6 @@ class GeoSeries(GeoFrame, pspd.Series):
         single_sided=False,
         **kwargs,
     ) -> "GeoSeries":
-        """
-        Returns a GeoSeries of geometries representing all points within a given distance of each geometric object.
-
-        Parameters
-        ----------
-        distance : float
-            The distance to buffer around each geometry.
-        resolution : int, optional, default 16
-            The resolution of the buffer around each geometry.
-        cap_style : str, optional, default "round"
-            The style of the buffer's cap (round, flat, or square).
-        join_style : str, optional, default "round"
-            The style of the buffer's join (round, mitre, or bevel).
-        mitre_limit : float, optional, default 5.0
-            The mitre limit for the buffer's join style.
-        single_sided : bool, optional, default False
-            Whether to create a single-sided buffer.
-
-        Returns
-        -------
-        GeoSeries
-            A GeoSeries of buffered geometries.
-        """
         spark_col = stf.ST_Buffer(self.spark.column, distance)
         return self._query_geometry_column(
             spark_col,
@@ -3300,64 +1477,6 @@ class GeoSeries(GeoFrame, pspd.Series):
         )
 
     def simplify(self, tolerance=None, preserve_topology=True) -> "GeoSeries":
-        """Returns a ``GeoSeries`` containing a simplified representation of
-        each geometry.
-
-        The algorithm (Douglas-Peucker) recursively splits the original line
-        into smaller parts and connects these parts' endpoints
-        by a straight line. Then, it removes all points whose distance
-        to the straight line is smaller than `tolerance`. It does not
-        move any points and it always preserves endpoints of
-        the original line or polygon.
-        See https://shapely.readthedocs.io/en/latest/manual.html#object.simplify
-        for details
-
-        Simplifies individual geometries independently, without considering
-        the topology of a potential polygonal coverage. If you would like to treat
-        the ``GeoSeries`` as a coverage and simplify its edges, while preserving the
-        coverage topology, see :meth:`simplify_coverage`.
-
-        Parameters
-        ----------
-        tolerance : float
-            All parts of a simplified geometry will be no more than
-            `tolerance` distance from the original. It has the same units
-            as the coordinate reference system of the GeoSeries.
-            For example, using `tolerance=100` in a projected CRS with meters
-            as units means a distance of 100 meters in reality.
-        preserve_topology: bool (default True)
-            False uses a quicker algorithm, but may produce self-intersecting
-            or otherwise invalid geometries.
-
-        Notes
-        -----
-        Invalid geometric objects may result from simplification that does not
-        preserve topology and simplification may be sensitive to the order of
-        coordinates: two geometries differing only in order of coordinates may be
-        simplified differently.
-
-        See also
-        --------
-        simplify_coverage : simplify geometries using coverage simplification
-
-        Examples
-        --------
-        >>> from sedona.geopandas import GeoSeries
-        >>> from shapely.geometry import Point, LineString
-        >>> s = GeoSeries(
-        ...     [Point(0, 0).buffer(1), LineString([(0, 0), (1, 10), (0, 20)])]
-        ... )
-        >>> s
-        0    POLYGON ((1 0, 0.99518 -0.09802, 0.98079 -0.19...
-        1                         LINESTRING (0 0, 1 10, 0 20)
-        dtype: geometry
-
-        >>> s.simplify(1)
-        0    POLYGON ((0 1, 0 -1, -1 0, 0 1))
-        1              LINESTRING (0 0, 0 20)
-        dtype: geometry
-        """
-
         spark_expr = (
             stf.ST_SimplifyPreserveTopology(self.spark.column, tolerance)
             if preserve_topology
@@ -3417,6 +1536,7 @@ class GeoSeries(GeoFrame, pspd.Series):
     def geometry(self) -> "GeoSeries":
         return self
 
+    # GeoSeries-only (not in GeoDataFrame)
     @property
     def x(self) -> pspd.Series:
         """Return the x location of point geometries in a GeoSeries
@@ -3450,6 +1570,7 @@ class GeoSeries(GeoFrame, pspd.Series):
             returns_geom=False,
         )
 
+    # GeoSeries-only (not in GeoDataFrame)
     @property
     def y(self) -> pspd.Series:
         """Return the y location of point geometries in a GeoSeries
@@ -3484,6 +1605,7 @@ class GeoSeries(GeoFrame, pspd.Series):
             returns_geom=False,
         )
 
+    # GeoSeries-only (not in GeoDataFrame)
     @property
     def z(self) -> pspd.Series:
         """Return the z location of point geometries in a GeoSeries
@@ -3518,6 +1640,7 @@ class GeoSeries(GeoFrame, pspd.Series):
             returns_geom=False,
         )
 
+    # GeoSeries-only (not in GeoDataFrame)
     @property
     def m(self) -> pspd.Series:
         raise NotImplementedError("GeoSeries.m() is not implemented yet.")
@@ -3526,6 +1649,7 @@ class GeoSeries(GeoFrame, pspd.Series):
     # CONSTRUCTION METHODS
     # ============================================================================
 
+    # GeoSeries-only (not in GeoDataFrame)
     @classmethod
     def from_file(
         cls, filename: str, format: Union[str, None] = None, **kwargs
@@ -3558,6 +1682,7 @@ class GeoSeries(GeoFrame, pspd.Series):
         df = sgpd.io.read_file(filename, format, **kwargs)
         return GeoSeries(df.geometry, crs=df.crs)
 
+    # GeoSeries-only (not in GeoDataFrame)
     @classmethod
     def from_wkb(
         cls,
@@ -3636,7 +1761,7 @@ class GeoSeries(GeoFrame, pspd.Series):
 
         schema = StructType([StructField("data", BinaryType(), True)])
         return cls._create_from_select(
-            f"ST_GeomFromWKB(`data`)",
+            stc.ST_GeomFromWKB(F.col("data")),
             data,
             schema,
             index,
@@ -3644,6 +1769,7 @@ class GeoSeries(GeoFrame, pspd.Series):
             **kwargs,
         )
 
+    # GeoSeries-only (not in GeoDataFrame)
     @classmethod
     def from_wkt(
         cls,
@@ -3715,7 +1841,7 @@ class GeoSeries(GeoFrame, pspd.Series):
 
         schema = StructType([StructField("data", StringType(), True)])
         return cls._create_from_select(
-            f"ST_GeomFromText(`data`)",
+            stc.ST_GeomFromText(F.col("data")),
             data,
             schema,
             index,
@@ -3723,6 +1849,7 @@ class GeoSeries(GeoFrame, pspd.Series):
             **kwargs,
         )
 
+    # GeoSeries-only (not in GeoDataFrame)
     @classmethod
     def from_xy(cls, x, y, z=None, index=None, crs=None, **kwargs) -> "GeoSeries":
         """
@@ -3781,11 +1908,11 @@ class GeoSeries(GeoFrame, pspd.Series):
 
         if z:
             data = list(zip(x, y, z))
-            select = f"ST_PointZ(`x`, `y`, `z`)"
+            select = stc.ST_PointZ(F.col("x"), F.col("y"), F.col("z"))
             schema.add(StructField("z", DoubleType(), True))
         else:
             data = list(zip(x, y))
-            select = f"ST_Point(`x`, `y`)"
+            select = stc.ST_Point(F.col("x"), F.col("y"))
 
         geoseries = cls._create_from_select(
             select,
@@ -3867,7 +1994,7 @@ class GeoSeries(GeoFrame, pspd.Series):
 
     @classmethod
     def _create_from_select(
-        cls, select: str, data, schema, index, crs, **kwargs
+        cls, spark_col: PySparkColumn, data, schema, index, crs, **kwargs
     ) -> "GeoSeries":
 
         from pyspark.pandas.utils import default_session
@@ -3877,7 +2004,7 @@ class GeoSeries(GeoFrame, pspd.Series):
         if isinstance(data, list) and not isinstance(data[0], (tuple, list)):
             data = [(obj,) for obj in data]
 
-        select = f"{select} as geometry"
+        name = kwargs.get("name", SPARK_DEFAULT_SERIES_NAME)
 
         if isinstance(data, pspd.Series):
             spark_df = data._internal.spark_frame
@@ -3888,29 +2015,26 @@ class GeoSeries(GeoFrame, pspd.Series):
         else:
             spark_df = default_session().createDataFrame(data, schema=schema)
 
-        spark_df = spark_df.selectExpr(select)
+        spark_df = spark_df.select(spark_col.alias(name))
 
         internal = InternalFrame(
             spark_frame=spark_df,
             index_spark_columns=None,
-            column_labels=[("geometry",)],
-            data_spark_columns=[scol_for(spark_df, "geometry")],
-            data_fields=[
-                InternalField(np.dtype("object"), spark_df.schema["geometry"])
-            ],
-            column_label_names=[("geometry",)],
+            column_labels=[(name,)],
+            data_spark_columns=[scol_for(spark_df, name)],
+            data_fields=[InternalField(np.dtype("object"), spark_df.schema[name])],
+            column_label_names=[(name,)],
         )
-        return GeoSeries(
-            first_series(PandasOnSparkDataFrame(internal)),
-            index,
-            crs=crs,
-            name=kwargs.get("name", None),
-        )
+        ps_series = first_series(PandasOnSparkDataFrame(internal))
+        name = None if name == SPARK_DEFAULT_SERIES_NAME else name
+        ps_series.rename(name, inplace=True)
+        return GeoSeries(ps_series, index, crs=crs)
 
     # ============================================================================
     # DATA ACCESS AND MANIPULATION
     # ============================================================================
 
+    # GeoSeries-only (not in GeoDataFrame)
     def isna(self) -> pspd.Series:
         """
         Detect missing values.
@@ -3950,12 +2074,14 @@ class GeoSeries(GeoFrame, pspd.Series):
             spark_expr,
             returns_geom=False,
         )
-        return to_bool(result)
+        return _to_bool(result)
 
+    # GeoSeries-only (not in GeoDataFrame)
     def isnull(self) -> pspd.Series:
         """Alias for `isna` method. See `isna` for more detail."""
         return self.isna()
 
+    # GeoSeries-only (not in GeoDataFrame)
     def notna(self) -> pspd.Series:
         """
         Detect non-missing values.
@@ -3996,12 +2122,14 @@ class GeoSeries(GeoFrame, pspd.Series):
             spark_expr,
             returns_geom=False,
         )
-        return to_bool(result)
+        return _to_bool(result)
 
+    # GeoSeries-only (not in GeoDataFrame)
     def notnull(self) -> pspd.Series:
         """Alias for `notna` method. See `notna` for more detail."""
         return self.notna()
 
+    # GeoSeries-only (not in GeoDataFrame)
     def fillna(
         self, value=None, inplace: bool = False, limit=None, **kwargs
     ) -> Union["GeoSeries", None]:
@@ -4013,8 +2141,8 @@ class GeoSeries(GeoFrame, pspd.Series):
         value : shapely geometry or GeoSeries, default None
             If None is passed, NA values will be filled with GEOMETRYCOLLECTION EMPTY.
             If a shapely geometry object is passed, it will be
-            used to fill all missing values. If a ``GeoSeries`` or ``GeometryArray``
-            are passed, missing values will be filled based on the corresponding index
+            used to fill all missing values. If a ``GeoSeries``
+            is passed, missing values will be filled based on the corresponding index
             locations. If pd.NA or np.nan are passed, values will be filled with
             ``None`` (not GEOMETRYCOLLECTION EMPTY).
         limit : int, default None
@@ -4080,7 +2208,6 @@ class GeoSeries(GeoFrame, pspd.Series):
         GeoSeries.isna : detect missing values
         """
         from shapely.geometry.base import BaseGeometry
-        from geopandas.array import GeometryArray
 
         # TODO: Implement limit https://github.com/apache/sedona/issues/2068
         if limit:
@@ -4104,7 +2231,7 @@ class GeoSeries(GeoFrame, pspd.Series):
             other, extended = self._make_series_of_val(value)
             align = False if extended else align
 
-        elif isinstance(value, (GeoSeries, GeometryArray, gpd.GeoSeries)):
+        elif isinstance(value, (GeoSeries, gpd.GeoSeries)):
 
             if not isinstance(value, GeoSeries):
                 value = GeoSeries(value)
@@ -4123,6 +2250,7 @@ class GeoSeries(GeoFrame, pspd.Series):
             align=align,
             returns_geom=True,
             default_val=None,
+            keep_name=True,
         )
 
         if inplace:
@@ -4242,33 +2370,6 @@ class GeoSeries(GeoFrame, pspd.Series):
 
     @property
     def bounds(self) -> pspd.DataFrame:
-        """Returns a ``DataFrame`` with columns ``minx``, ``miny``, ``maxx``,
-        ``maxy`` values containing the bounds for each geometry.
-
-        See ``GeoSeries.total_bounds`` for the limits of the entire series.
-
-        Examples
-        --------
-        >>> from shapely.geometry import Point, Polygon, LineString
-        >>> d = {'geometry': [Point(2, 1), Polygon([(0, 0), (1, 1), (1, 0)]),
-        ... LineString([(0, 1), (1, 2)])]}
-        >>> gdf = geopandas.GeoDataFrame(d, crs="EPSG:4326")
-        >>> gdf.bounds
-           minx  miny  maxx  maxy
-        0   2.0   1.0   2.0   1.0
-        1   0.0   0.0   1.0   1.0
-        2   0.0   1.0   1.0   2.0
-
-        You can assign the bounds to the ``GeoDataFrame`` as:
-
-        >>> import pandas as pd
-        >>> gdf = pd.concat([gdf, gdf.bounds], axis=1)
-        >>> gdf
-                                geometry  minx  miny  maxx  maxy
-        0                     POINT (2 1)   2.0   1.0   2.0   1.0
-        1  POLYGON ((0 0, 1 1, 1 0, 0 0))   0.0   0.0   1.0   1.0
-        2           LINESTRING (0 1, 1 2)   0.0   1.0   1.0   2.0
-        """
         selects = [
             stf.ST_XMin(self.spark.column).alias("minx"),
             stf.ST_YMin(self.spark.column).alias("miny"),
@@ -4295,21 +2396,6 @@ class GeoSeries(GeoFrame, pspd.Series):
 
     @property
     def total_bounds(self):
-        """Returns a tuple containing ``minx``, ``miny``, ``maxx``, ``maxy``
-        values for the bounds of the series as a whole.
-
-        See ``GeoSeries.bounds`` for the bounds of the geometries contained in
-        the series.
-
-        Examples
-        --------
-        >>> from shapely.geometry import Point, Polygon, LineString
-        >>> d = {'geometry': [Point(3, -1), Polygon([(0, 0), (1, 1), (1, 0)]),
-        ... LineString([(0, 1), (1, 2)])]}
-        >>> gdf = geopandas.GeoDataFrame(d, crs="EPSG:4326")
-        >>> gdf.total_bounds
-        array([ 0., -1.,  3.,  2.])
-        """
         import numpy as np
         import warnings
         from pyspark.sql import functions as F
@@ -4342,12 +2428,9 @@ class GeoSeries(GeoFrame, pspd.Series):
                 )
             )
 
+    # GeoSeries-only (not in GeoDataFrame)
     def estimate_utm_crs(self, datum_name: str = "WGS 84") -> "CRS":
         """Returns the estimated UTM CRS based on the bounds of the dataset.
-
-        .. versionadded:: 0.9
-
-        .. note:: Requires pyproj 3+
 
         Parameters
         ----------
@@ -4493,7 +2576,7 @@ e": "Feature", "properties": {}, "geometry": {"type": "Point", "coordinates": [3
         --------
         GeoSeries.to_file : write GeoSeries to file
         """
-        return self._to_geoframe(name="geometry").to_json(
+        return self.to_geoframe(name="geometry").to_json(
             na="null", show_bbox=show_bbox, drop_id=drop_id, to_wgs84=to_wgs84, **kwargs
         )
 
@@ -4747,7 +2830,7 @@ e": "Feature", "properties": {}, "geometry": {"type": "Point", "coordinates": [3
         >>> gdf = GeoDataFrame({"geometry": [Point(0, 0)]}, index=["a", "b"])
         >>> gdf.to_file(gdf, driver="geoparquet")
         """
-        self._to_geoframe().to_file(path, driver, index=index, **kwargs)
+        self.to_geoframe().to_file(path, driver, index=index, **kwargs)
 
     def to_parquet(self, path, **kwargs):
         """
@@ -4758,7 +2841,7 @@ e": "Feature", "properties": {}, "geometry": {"type": "Point", "coordinates": [3
         - kwargs: Any
             Additional arguments to pass to the Sedona DataFrame output function.
         """
-        self._to_geoframe().to_file(path, driver="geoparquet", **kwargs)
+        self.to_geoframe().to_file(path, driver="geoparquet", **kwargs)
 
     # -----------------------------------------------------------------------------
     # # Utils
@@ -4787,10 +2870,10 @@ e": "Feature", "properties": {}, "geometry": {"type": "Point", "coordinates": [3
         else:
             return value, False
 
-    def _to_geoframe(self, name=None):
+    def to_geoframe(self, name=None):
         if name is not None:
             renamed = self.rename(name)
-        elif self._column_label is None:
+        elif self._column_label is None or self._column_label == (None,):
             renamed = self.rename("geometry")
         else:
             renamed = self
@@ -4808,7 +2891,7 @@ def _get_series_col_name(ps_series: pspd.Series) -> str:
     return ps_series.name if ps_series.name else SPARK_DEFAULT_SERIES_NAME
 
 
-def to_bool(ps_series: pspd.Series, default: bool = False) -> pspd.Series:
+def _to_bool(ps_series: pspd.Series, default: bool = False) -> pspd.Series:
     """
     Cast a ps.Series to bool type if it's not one, converting None values to the default value.
     """
@@ -4817,16 +2900,3 @@ def to_bool(ps_series: pspd.Series, default: bool = False) -> pspd.Series:
         ps_series.fillna(default, inplace=True)
 
     return ps_series
-
-
-def _to_geo_series(df: PandasOnSparkSeries) -> GeoSeries:
-    """
-    Get the first Series from the DataFrame.
-
-    Parameters:
-    - df: The input DataFrame.
-
-    Returns:
-    - GeoSeries: The first Series from the DataFrame.
-    """
-    return GeoSeries(data=df)
